Create overflow_example.c
This what written to help me figure out how the variables are stored in memory. It includes every memory address involved for each variable & the value assigned to it. 

This allowed me to dynamically put in data and see what it's doing to the memory addresses & how it flows into the next buffer. 

I discovered several interesting things that's vital to learning buffer overflows: 

First when a string is written to memory it ends each string with a null character '\0'. So if you're trying to write 8 characters to an 8-byte buffer then it will overflow because that trailing null byte will overwrite the next buffer. So with an 8-byte buffer you can only write 7 characters & then leave that last byte for the null byte, this will safely write to an 8-byte buffer. Also when a computer then tries to read a variable that's overflowed, it will keep reading the string until it hits that null byte, no mater if it goes past the 8 bytes that's allocated for that variable. 

Secondly I learned about little-endian format when data is written to memory. That's important because the buffer will only overflow into memory address higher in value than the source variable that's being written to. So it can overwrite variables that are higher up in memory but not below it.